added validation code and some code fixes

+++ Plugins/DeployCommand/DeployCommand.cs
@@ -4,12 +4,12 @@
 // http://www.codefarts.com
 // </copyright>
 
+using System;
+
 namespace Codefarts.BuildHelper
 {
-    using System;
     using System.IO;
     using System.Linq;
-    using System.Xml;
 
     public class DeployCommand : IBuildCommand
     {
@@ -17,43 +17,29 @@ namespace Codefarts.BuildHelper
 
         public void Execute(ExecuteCommandArgs args)
         {
-            var value = args.Element.GetValue("path");
-            var destPath = value != null ? value.ReplaceBuildVariableStrings(args.Variables) : null;
+            if (args == null)
+            {
+                throw new ArgumentNullException(nameof(args));
+            }
+
+            if (args.Variables == null)
+            {
+                throw new BuildException("Variables dictionary is null.");
+            }
+
+            var destPath = args.GetParameter<string>("path", null);
+            destPath = destPath != null ? destPath.ReplaceBuildVariableStrings(args.Variables) : null;
             if (destPath == null)
             {
-                throw new XmlException($"Command: {nameof(DeployCommand)} value: path  - Value not found");
+                throw new BuildException($"Command: {nameof(DeployCommand)} value: path  - Value not found");
             }
-
-            var message = args.Element.GetValue("message");
 
             // check if we should clear the folder first
-            value = args.Element.GetValue("clean");
-            var doClear = string.IsNullOrWhiteSpace(value) ? false : value.Trim().ToLowerInvariant() == "true";
+            var doClean = args.GetParameter("clean", false);
 
-            // check type of conditions
-            var conditionsValue = args.Element.GetValue("allconditions");
-            var allConditions = true;
-            if (conditionsValue != null && !bool.TryParse(conditionsValue, out allConditions))
-            {
-                throw new ArgumentOutOfRangeException($"'{allConditions}' attribute exists but it's value could not be parsed as a bool value.");
-            }
-
-            // check conditions
-            if (!args.Element.SatifiesConditions(args.Variables, allConditions))
-            {
-                args.Output("Conditions not satisfied.");
-                return;
-            }
-
-            if (!string.IsNullOrWhiteSpace(message))
-            {
-                message = message.ReplaceBuildVariableStrings(args.Variables);
-                args.Output($"Message: {message}");
-            }
-
-            args.Output($"Clearing before deploy ({doClear}): {destPath}");
+            args.Output($"Clearing before deploy ({doClean}): {destPath}");
             var di = new DirectoryInfo(destPath);
-            if (doClear && di.Exists)
+            if (doClean && di.Exists)
             {
                 foreach (var file in di.EnumerateFiles())
                 {
@@ -66,16 +52,34 @@ namespace Codefarts.BuildHelper
                 }
             }
 
-            var srcPath = "$(ProjectDir)$(OutDir)".ReplaceBuildVariableStrings(args.Variables);
-            var allFiles = Directory.GetFiles(srcPath, "*.*", SearchOption.AllDirectories).Select(d => Path.GetFileName(d));
-            foreach (var file in allFiles)
+            // ensure there is a ProjectDir and OutDir variables
+            if (!args.Variables.ContainsKey("ProjectDir"))
             {
-                var src = Path.Combine(srcPath, file);
-                var dest = Path.Combine(destPath, file);
+                throw new BuildException("Deploy command requires a 'ProjectDir' variable to run.");
+            }
 
-                Directory.CreateDirectory(Path.GetDirectoryName(dest));
-                args.Output("Deploying: " + src + " ==> " + dest);
-                File.Copy(src, dest, true);
+            if (!args.Variables.ContainsKey("OutDir"))
+            {
+                throw new BuildException("Deploy command requires a 'OutDir' variable to run.");
+            }
+
+            var srcPath = Path.Combine("$(ProjectDir)".ReplaceBuildVariableStrings(args.Variables),
+                                       "$(OutDir)".ReplaceBuildVariableStrings(args.Variables));
+            if (Directory.Exists(srcPath))
+            {
+                var allFiles = Directory.GetFiles(srcPath, "*.*", SearchOption.AllDirectories);//.Select(d => Path.GetFileName(d));
+                foreach (var file in allFiles)
+                {
+                    var filePath = file.Substring(srcPath.Length).Trim();
+                    filePath = filePath.TrimStart(Path.DirectorySeparatorChar, Path.AltDirectorySeparatorChar);
+
+                    var src = Path.Combine(srcPath, filePath);
+                    var dest = Path.Combine(destPath, filePath);
+
+                    Directory.CreateDirectory(Path.GetDirectoryName(dest));
+                    args.Output("Deploying: " + src + " ==> " + dest);
+                    File.Copy(src, dest, true);
+                }
             }
         }
     }
